Fix for PropertyConverter.
Was converting valuetype long to integer.

+++ eclipseProjects/de.enflexit.awb/bundles/web/de.enflexit.awb.ws.restapi/src/de/enflexit/awb/ws/restapi/tools/PropertyConverter.java
@@ -92,7 +92,7 @@ public class PropertyConverter {
 			pEntry.setValueType(ValueType.INTEGER);
 			break;
 		case Long:
-			pEntry.setValueType(ValueType.INTEGER);
+			pEntry.setValueType(ValueType.LONG);
 			break;
 		}
 		return pEntry;
